# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Cuarto Arte API - Gestión de eventos musicales y artísticos
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /clients:
    get:
      operationId: getClients
      tags: [clients]
      summary: List all clients
      responses:
        "200":
          description: List of clients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Client"
    post:
      operationId: createClient
      tags: [clients]
      summary: Create a new client
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClient"
      responses:
        "201":
          description: Client created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
  /clients/{id}:
    get:
      operationId: getClient
      tags: [clients]
      summary: Get a client by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Client found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
        "404":
          description: Client not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateClient
      tags: [clients]
      summary: Update a client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClient"
      responses:
        "200":
          description: Client updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
    delete:
      operationId: deleteClient
      tags: [clients]
      summary: Delete a client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Client deleted
...
  /events:
    get:
      operationId: getEvents
      tags: [events]
      summary: List all events
      parameters:
        - name: startDate
          in: query
          schema:
            type: string
        - name: endDate
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Event"
    post:
      operationId: createEvent
      tags: [events]
      summary: Create a new event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEvent"
      responses:
        "201":
          description: Event created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
  /events/{id}:
    get:
      operationId: getEvent
      tags: [events]
      summary: Get an event by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Event found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateEvent
      tags: [events]
      summary: Update an event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEvent"
      responses:
        "200":
          description: Event updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
    delete:
      operationId: deleteEvent
      tags: [events]
      summary: Delete an event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Event deleted
  /events/{id}/musicians:
    get:
      operationId: getEventMusicians
      tags: [events]
      summary: Get musicians assigned to an event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Event musicians
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EventMusician"
    post:
      operationId: assignMusiciansToEvent
      tags: [events]
      summary: Assign musicians to an event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                musicianIds:
                  type: array
                  items:
                    type: integer
              required:
                - musicianIds
      responses:
        "200":
          description: Musicians assigned
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EventMusician"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    Client:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        email:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - name
        - phone
        - createdAt
    CreateClient:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        email:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
      required:
        - name
        - phone
...
    Event:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        clientId:
          type: integer
          nullable: true
        clientName:
          type: string
          nullable: true
        date:
          type: string
        time:
          type: string
        venue:
          type: string
        eventType:
          type: string
        status:
          type: string
          enum: [pendiente, confirmado, completado, cancelado]
        totalAmount:
          type: number
        advanceAmount:
          type: number
        notes:
          type: string
          nullable: true
        musicians:
          type: array
          items:
            $ref: "#/components/schemas/EventMusician"
        createdAt:
          type: string
      required:
        - id
        - title
        - date
        - time
        - venue
        - eventType
        - status
        - totalAmount
        - advanceAmount
        - createdAt
    CreateEvent:
      type: object
      properties:
        title:
          type: string
        clientId:
          type: integer
          nullable: true
        date:
          type: string
        time:
          type: string
        venue:
          type: string
        eventType:
          type: string
        status:
          type: string
          enum: [pendiente, confirmado, completado, cancelado]
        totalAmount:
          type: number
        advanceAmount:
          type: number
        notes:
          type: string
          nullable: true
        musicianIds:
          type: array
          items:
            type: integer
      required:
        - title
        - date
        - time
        - venue
        - eventType
        - status
        - totalAmount
        - advanceAmount
    EventMusician:
      type: object
      properties:
        id:
          type: integer
        eventId:
          type: integer
        musicianId:
          type: integer
        musicianName:
          type: string
        instruments:
          type: string
        fee:
          type: number
          nullable: true
      required:
        - id
        - eventId
        - musicianId
        - musicianName
        - instruments
...