# lib/api-zod/src/generated/types/createEvent.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Cuarto Arte API - Gestión de eventos musicales y artísticos
 * OpenAPI spec version: 0.1.0
 */
import type { CreateEventStatus } from "./createEventStatus";

export interface CreateEvent {
  title: string;
  clientId?: number | null;
  date: string;
  time: string;
  venue: string;
  eventType: string;
  status: CreateEventStatus;
  totalAmount: number;
  advanceAmount: number;
  notes?: string | null;
  musicianIds?: number[];
}