# lib/api-zod/src/generated/types/event.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Cuarto Arte API - Gestión de eventos musicales y artísticos
 * OpenAPI spec version: 0.1.0
 */
import type { EventMusician } from "./eventMusician";
import type { EventStatus } from "./eventStatus";

export interface Event {
  id: number;
  title: string;
  clientId?: number | null;
  clientName?: string | null;
  date: string;
  time: string;
  venue: string;
  eventType: string;
  status: EventStatus;
  totalAmount: number;
  advanceAmount: number;
  notes?: string | null;
  musicians?: EventMusician[];
  createdAt: string;
}