# lib/api-zod/src/generated/types/eventMusician.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Cuarto Arte API - Gestión de eventos musicales y artísticos
 * OpenAPI spec version: 0.1.0
 */

export interface EventMusician {
  id: number;
  eventId: number;
  musicianId: number;
  musicianName: string;
  instruments: string;
  fee?: number | null;
}